+++ src/clj/withings_client/middleware.clj
@@ -1,16 +1,19 @@
 (ns withings-client.middleware
   (:require
-   [withings-client.env :refer [defaults]]
+   [buddy.auth.middleware :refer [wrap-authentication wrap-authorization]]
+   [buddy.auth :refer [authenticated?]]
+   [buddy.auth.accessrules :refer [restrict]]
+   [buddy.auth.backends.session :refer [session-backend]]
    [clojure.tools.logging :as log]
-   [withings-client.layout :refer [error-page]]
-   [ring.middleware.anti-forgery :refer [wrap-anti-forgery]]
-   [withings-client.middleware.formats :as formats]
    [muuntaja.middleware :refer [wrap-format wrap-params]]
-   [withings-client.config :refer [env]]
-   [ring.middleware.flash :refer [wrap-flash]]
    [ring.adapter.undertow.middleware.session :refer [wrap-session]]
-   [ring.middleware.defaults :refer [site-defaults wrap-defaults]]))
-
+   [ring.middleware.anti-forgery :refer [wrap-anti-forgery]]
+   [ring.middleware.defaults :refer [site-defaults wrap-defaults]]
+   [ring.middleware.flash :refer [wrap-flash]]
+   [withings-client.config :refer [env]]
+   [withings-client.env :refer [defaults]]
+   [withings-client.layout :refer [error-page]]
+   [withings-client.middleware.formats :as formats]))
 
 (defn wrap-internal-error [handler]
   (fn [req]
@@ -30,7 +33,6 @@
      {:status 403
       :title "Invalid anti-forgery token"})}))
 
-
 (defn wrap-formats [handler]
   (let [wrapped (-> handler wrap-params (wrap-format formats/instance))]
     (fn [request]
@@ -38,8 +40,26 @@
       ;; since they're not compatible with this middleware
       ((if (:websocket? request) handler wrapped) request))))
 
+;; restriction
+(defn on-error [request response]
+  (error-page
+   {:status 403
+    :title (str "Access to " (:uri request) " is not authorized")}))
+
+(defn wrap-restricted [handler]
+  (restrict handler {:handler authenticated?
+                     :on-error on-error}))
+
+(defn wrap-auth [handler]
+  (let [backend (session-backend)]
+    (-> handler
+        (wrap-authentication backend)
+        (wrap-authorization backend))))
+;; restriction end
+
 (defn wrap-base [handler]
   (-> ((:middleware defaults) handler)
+      wrap-auth ;; restriction
       wrap-flash
       (wrap-session {:cookie-attrs {:http-only true}})
       (wrap-defaults